explode api to a single file for each
